Tech debt cleanup
Add cleanup where the profile was not updating the correct time
when the appearance settings were changed. Updated the db layer
to handle the update of timestamp with time zone for the user modal.

Added cleanup where sharing 1 category with a third party was not allowing
the user to see that the item was being shared. rather the user had to
refresh the page. Fixed this issue so that the user does not have to
refresh the page anymore.
