+++ frontend/src/features/MaintenancePlanItemDetails/maintenancePlanItemSlice.js
@@ -27,6 +27,24 @@ const maintenancePlanItemSlice = createSlice({
       state.error = '';
       state.selectedMaintenancePlan = {};
     },
+    updatePlan: (state) => {
+      state.loading = true;
+      state.error = '';
+    },
+    updatePlanSuccess: (state, action) => {
+      const draftMaintenancePlan = action.payload;
+      const filteredMaintenancePlan = [...state.maintenancePlan].filter((v) => {
+        return v.id !== draftMaintenancePlan.id;
+      });
+      state.maintenancePlan = [draftMaintenancePlan, ...filteredMaintenancePlan];
+      state.selectedMaintenancePlan = draftMaintenancePlan;
+      state.loading = false;
+      state.error = '';
+    },
+    updatePlanFailure: (state) => {
+      state.loading = false;
+      state.error = '';
+    },
     getItemsInMaintenancePlan: (state) => {
       state.loading = true;
       state.error = '';
